Use PROCD_ADDRESS default in old Starter
Previously, PROCD_ADDRESS had a default value everywhere
but the old Starter. The old Starter only needs to look up
PROCD_ADDRESS if PrivSep is on. So this fix allows standard
universe jobs to work on pools using PrivSep when
PROCD_ADDRESS is not explicitly set (which it is in our
default config file).