Reorder GlobalJobId to be SCHEDD_NAME#ClusterId.ProcId#QDate, as per 24 June 2008 FW meeting
Idea is to make the GlobalJobId fields, which should not be parsed
(OPAQUE!), go left to right as slowest changing to fastest
changing. This assists in using regexp to find jobs when you only want
to keep track of part of the Id, say everything except the QDate.

NOTE: The last field isn't really the QDate, it's just very close to
the QDate.