Eliminate partial writes for DaemonCore::Write_Pipe() on Windows.
On unix, a write() on a pipe will block until all the data is written.
Until now, the DaemonCore pipe code on Windows could return after
writing only some of the data. This broke the C-GAHP, which assumed the
unix convention. The result is that the gridmanager would EXCEPT() when
the C-GAHP returned invalid data.

Now, the Windows pipe code in DaemonCore will block on writes until the
full data buffer is written.