Added the ability for DaemonCore to update the timestamps of valid lock files at periodic intervals.
Inside the FileLock object there is now a static singly linked list in
which is stored FileLockEntry structures for FileLock objects which have a
valid path component (now required when the fd or fp is valid). An object
may only be in the list if it has a valid path component. Each object may
update its own timestamp, or a static member function may iterate across
the list and update them--called by daemoncore at regular intervals. The
intervals are described by LOCK_FILE_UPDATE_INTERVAL in the config file.

All locks are updated with utime() and _only in the condor privstate_. This
catches things like the InstanceLock, but not the userlog. For now this is
good enough, and it is pretty difficult to decide how to handle the
permission problem with lock files from arbitrary users.

I had to write a linked list implementation manually here due to the fact that
file_lock.h is included in libcondorapi. This means that if I brought in
say, HashTable, I'd have to bring in a wealth of other header files and that's
just painful.

This was code reviewed by gthain and jfrey.