Change condor_findhost's subsystem to TOOL
Tools get a session duration of 1 minute in CEDAR. Everything else
(other than condor_submit) get a session duration of 100 days. So with
condor_findhost's old subsystem of INTERACTIVE, many invocations will
bloat the collector with useless sessions. There's no reason for
condor_findhost to have it's own subsystem, so just change it to TOOL
and let it get the shorter sessions.