Avoid excessive proxy refreshes for gt2 jobs
Refreshing the proxy of a gt2 gram job requires starting and killing a
jobmanager on the remote system. We've seen sites with thousands of jobs
update their proxies every half hour, with a new lifetime of 4 days.
That generates a lot of unnecessary work.

Now, when the gridmanager gets a new proxy for a gt2 job, it forwards it
only when the job's current proxy expires in less than 6 hours.