Fix potential discarding of job status updates for gt2 jobs
If a job status update of DONE (which is turned into STAGE_OUT) arrives
from the grid monitor while the gridmanager is trying to restart the
jobmanager, the update can be lost. Once the restart completes and the
GlobusJob moves to GM_SUBMITTED state, it decides whether to put the
jobmanager back to sleep before checking if there are any unconsumed job
status callbacks. Then, when the gridmanager tells the jobmanager to
exit, the jobmanager replies with a new callback saying it's going away,
which overwrites the previous callback of DONE.

Now, we check for pending callbacks in GM_SUBMITTED before deciding
whether the jobmanager should be put to sleep.