Add support for GRAM 4.2, with auto-distinguishing between it and GRAM 4.0
The protocol for WS GRAM is slightly different between Globus 4.0 and
4.2. Apparently, the underlying WS Addressing standard changed. Now,
Condor supports GRAM 4.2 in addition to GRAM 4.0.

We're doing our best to hide the difference from users. So while there's
a new GT42 GAHP, the gridmanager still has a single gt4 grid-type. The
GT4Job and GT4Resource classes will first attempt to use the GT4 GAHP to
speak with the resource. If they get an error indicating they're talking
to a GRAM 4.2 resource, they'll automatically switch to a GT42 GAHP.