Protect schedd against spending too much time negotiating for stale matches. The problem that has been observed is with match info from negotiator to startds turned off, a schedd that is behind on requesting claims can end up getting very little useful work done, because it spends too much time negotiating and getting matches to machines that it already has in its queue of startds to contact.
A more correct solution to this problem requires the negotiator to keep some
state, and, in fact, this is already possible.  However, nobody configures
things that way by default, so the schedd should protect itself when it can.
Also removed the match_rec state M_CONNECTING, since it was never entering
that state.