Implemented configuration for Concurrency Limits, %s_LIMIT and CONCURRENCY_LIMIT_DEFAULT
All limits can be bound via X_LIMIT in a configuration file, where X
is the case-insensitive name of the limit, e.g. MATLAB_LIMIT = 100
allows for 100 jobs requiring MATLAB to run at one time. A
CONCURRENCY_LIMIT_DEFAULT is also provided to handle when X_LIMIT does
not exist. The default's default is sufficiently large to not limit
much of anything.