Fixed two issues, 1) qmgmt.C must strlwr ConcurrencyLimits, 2) if both job and match didn't use ConcurrencyLimits the match would not be recycled
The first issue is to enforce case-insensitive of limits and handle
the case where "X" and "x" are given as limits. Previously
matchmaker.C would strlwr, but if qmgmt.C did not then claim reuse
would not happen for "X" and "x" when it should. The
case-insensitivity is required because limit bounds are configuration
variables and configuration variable names are case-insensitive.

The second issue was that if CLAIM_RECYCLING_CONSIDER_LIMITS was TRUE
then lack of a ConcurrencyLimits attribute on either a job or machine
ad would result in recycling being skipped. This is a problem the
claim should be recycled if the attribute is missing from both.