Record preempted matches in the Accountant, when they reappear at least
The Accountant can only have one match to a given resource at a
time. Matches are keyed on resource names. When preemption occurs a
match already exists in the Accountant with the same name as the
preempting match. The preempted (original) match is immediately
discarded and the preempting (new) match is recorded. Now, if
preemption and claiming by the new match takes longer than a
negotiation cycle, or Startd advertisements are dropped, the
new match is also discarded. The result is the Accountant has no
record of any match associated with the resource until the next
successful, claimed|matched, startd advertizement arrives for a
negotiation cycle.

This is the result of Accountant::CheckClaimedOrMatched and
Accountant::IsClaimed not considering the Preempting state. This also
means that no customer (user) is charged for time during which their
job is preempting.

The fix in this patch is quite simple, it considers the
Preempting state to be equivalent to the Claimed state. This means
when a Preempting Startd advertisement arrived it will be re-added to
the set of known matches in the Accountant.