New approach, remember but gather, because of issues with previous approach, remember and store
The previous approach recorded limits in use by matches and stored
that information in the Accountant's log. The problem was that matches
are too volatile in the face of preemption and stale Startd ads.

The new approach is to remember limits as they are handed out, but on
each negotiation cycle to throw out the presumed view of limits and
gather the information reported from Startd ads. This approach does
not fall victim to instability caused by preemption. However, it still
falls victim to stale Startd ads.

The previous approach's issue with stale Startd ads was mitigated by
the MATCH_INFO, async and UDP, command from the Negotiator ->
Startd. The Startds with limits associated with them needed to move to
at least the Matched state between negotiation cycles.

This is different for the new approach. If a match is handed out in
one negotiation cycle and not used, i.e. Startd informed of the match
by the Schedd, by the following negotiation cycle, then the Accountant
may throw out the match until a negotiation cycle occurs where the
match is recreated from Startd ad information. The MATCH_INFO command
does not help here because a Matched Startd ad does not contain a
record of limits.

During the time the match is absent from the Accountant's memory, the
limits associated with that match will be under represented and may be
oversubscribed.

The ultimate solution is to have leases on matches. This would also
assist in more accurate accounting information.

A potential optimization for the new approach is to add "store" back
to the mix and read stored limits from Startd ads that are
Matched. This effectively would bring the stale window back in line
with the previous approach, and still be resilient in the case of
preemption.

Keep in mind that the "store" aspect is no help in the face of
preemption, and therefore the all approaches are vulnerable to the
Schedd->Startd claim race with negotiation cycles in the case of
preemption.