New hybrid approach, remember + store but gather
This is a hybrid approach that stores limit information in the
Accountant's log for use when Matched Startd ads arrive without their
own limit information.

This brings the "new" approach in line with the "previous" approach
with respect to stale Startd ad issues.

In the presence of preemption, assuming the Negotiator cannot hand out
a single ClaimId multiple times in a single negotiation cycle, the
existing hole is if a negotiation cycle happens after a preemption
match is created and before the Startd sends an update with
information of the preempting limits. In this case the preempting
limits will be forgotten.

It may still be possible to address this hole by merging the
Accontant's log information with the incoming Startd ads. However,
this is an error prone solution and may be trying to milk a rock.