Fold gahp and nordugrid_gahp externals into the Condor source.
The GT2 and NorduGrid GAHP servers are developed as part of Condor, but
have been treated as separate codebases as we envisioned them being used
by third parties apart from Condor. This has turned out not be the case,
so we're folding them into the regular Condor codebase. This will
greatly simplify their maintainence.

Until now, this code resided in the old Condor CVS repository. The
initial revision checked in here is a direct copy, except for the
makefiles.