Change Globus external to 4.2.0 and add OpenSSL extern 0.9.8h
Now that Globus will build against an external version of OpenSSL, let's
use a modern OpenSSL. For now, we don't build against the system OpenSSL
for two reasons:
* We can't do completely static builds, since the system OpenSSL on some
  systems need to link with Kerberos, which only provides shared
  libraries.
* The SysCall library pulls a subset of object files out of the OpenSSL
  libraries. The list of necessary files varies with the version of
  OpenSSL, making it a pain to build if we can't expect a specific
  version of OpenSSL.

The new Globus 4.2.0 external includes the GRAM4 and GRAM42 jar files,
allowing us to interalize the GT4 and GT42 GAHPs. For now, we check in
both the source code and the jars for the GAHPs.

The GT2 and NorduGrid GAHPs need to change because they modify some of
Globus's internal code.