Patched overflow in lookup_macro
The schedd will param for user supplied data, such as
SCHEDD_ROUND_ATTR_<attr_name>, meaning a user has the ability to
overflow the 1024 byte buffer in lookup_macro.

This fix changes the strcpy to a strncpy, which effectively truncates
the name of the macro that is being looked up. Theoretically the 1023
byte limit could be removed from macro names, but this fix is simple,
strictly better than was existed and does not reduce any existing
functionality.

This bug has existed since Tue Apr 4 21:50:18 2006 +0000.

735P 8E8X V7J4 6EWJ