Claimids are now encrypted over the network, when possible.
Fixed: Toggling of UDP encryption state did not work.  Previously, it had
to be turned on/off for the whole session and could not be toggled
for portions of the message.

Fixed: Toggling of TCP encryption state did not work for inherited sockets
because the encryption state was not getting properly reset at the
end of each message and this state was not passed to the child.