Fix ProcD crash caught by nightly tests
The problem here was that the ProcD would look up one of
its subfamilies by PID, then do a snapshot, then operate
on the subfamily. This is bad since the snapshot could
result in the looked-up subfamily being deleted, which
would cause a crash. We now do the snapshot defore the
lookup.