Fixed a bug affecting all calls to exit() in the child branch of CreateProcess(). The problem is that in 7.1.2, DaemonCore's exit() replacement was changed to call the real exit() rather than _exit(), so the parent's memory was getting clobbered if we were using clone().
This also fixes a longer standing problem in which calls to EXCEPT
or exit() inside the child branch of the code (i.e. utility functions
called by that code, such as dprintf) did not write an error
to the error pipe, so the parent did not know that exec() failed.