Handle fractional disk issue
The disk usage is stored as a fraction (a float) of the total
available disk space. When creating a slot that has just as much disk
as is requested the math, with the help of a floor(), can result in a
slot that has too little disk space. The solution here is to add 1KB
to all incoming disk requests to ensure the dynamic slot has enough
disk space, with at most 1KB more than requested.