Handle recycling issue, always include IsStillPartitionable
The schedd may recycle a claim, and in doing so it does not know the
actual requirements of the dynamic slot, only of the partitionable
slot. The result is the schedd can reuse a claim for jobs that are too
big for the dynamic slot. Making every slot check cpu/memory/disk
allows for the startd to reject such reuses.

It would be nice if the schedd could handle this as well - optimization.

NOTE: One possible issue, a dynamic slot can be reused for jobs that
are smaller than the slot - a DYNAMIC_SLOT expression could make sure
that jobs fix the slot exactly (with 1KB range for disk, of course)