This patch makes a backup of the job queue log in the event of certain failures.
Specifically, it can keep a backup of the job queue log on a local
filesystem in case writes, flushes, or syncs to the job queue log (on,
e.g., NFS) fail. Such a backup enables an administrator to see that a
job failed to submit, but does not perform any automatic recovery.

Here's how it works:

1. If there is a LOCAL_QUEUE_BACKUP_DIR Condor parameter that contains a
valid directory name, then we will create a local backup file in that
directory with a unique name. In this case, we will assume that we
might want a backup of the job queue log. (If not, we will not make a
local backup; proceed to #4.)

2. Every operation we do to the job queue log: writes, flushes, and
syncs, we will also do to the local file. We will track the first
error we see writing to the job queue log and will stop writing to the
actual job queue log -- but not EXCEPT -- when we encounter the first
error.

3. Once we've completed writing log records, we will unlink the local
backup if there were no errors writing the actual job queue log,
*unless* the parameter LOCAL_XACT_BACKUP_FILTER is set to "ALL".

4. If there was a failure writing the actual job queue log, we will
then report (via EXCEPT) the first error we observed while writing the
actual job queue log, as well as the location of the local backup
file, if we made one.

Note that we only sync the local (backup) file if we will be keeping
it, so the performance impact should be minimal in the common case.

This patch adds the following Condor parameters:

* LOCAL_QUEUE_BACKUP_DIR

LOCAL_QUEUE_BACKUP_DIR should be a directory on a non-network filesystem.

* LOCAL_XACT_BACKUP_FILTER

The value of the LOCAL_XACT_BACKUP_FILTER parameter filters whether or
not to backup any transactions based on whether or not the commit was
successful. Its value should be one of the following:

  "ALL", if we always want to keep local xact backups, no matter what,
  "NONE", if we should never keep local xact backups, or
  "FAILED", if we should only keep backups of xacts that failed to commit

The user must set LOCAL_QUEUE_BACKUP_DIR to a valid directory and
LOCAL_XACT_BACKUP_FILTER to something other than "NONE" in order to
retain any backups. A LOCAL_XACT_BACKUP_FILTER value of "ALL" is
likely only useful for Condor developers.