Add 'job status unknown' state to grid universe jobs
If too much time elapses without an update on a job's status from the
remote site, CurrentStatusUnknown is set to True in the job ad and an
event is written to the user log. When an update subsequently arrives,
CurrentStatusUnknown is set to False and a corresponding event is
written to the user log.

Also, LastRemoteStatusUpdate is set to the last time a job status
update arrived.