Added defaults for High Availability logs in condor_config.generic
Anything that dprintfs may EXCEPT if its log is not available, or
configured. For instance, I added ALL_DEBUG = D_FULLDEBUG and all of a
sudden the condor_transferer started EXCEPT'ing. This was because
TRANSFERER_LOG was not defined. It seems that a sane default for *_LOG
options should be $(LOG)/[SUBSYS]Log. This could also simplify
the config file.