Poor fix for condor.init restart
condor_master removes its pidfile before exiting, but a fast shutdown
can take some time anyway. During a restart (a quick stop then start)
the pidfile will exist when the start begins, if the master hasn't
exited. The result is start will fail and the master will not be
restarted.

A proper fix for this would be to watch the master's pid until it
finally exists and then say shutdown was successful, but that could
take forever and the init script should not hang.

So, the current alternative is to give the shutdown a few seconds to
complete, then proceed. This solution is no worse than what happened
when killproc was not provided a signal. It's actually strictly better
since SIGKILL is not sent to the condor_master. However, if the fast
shutdown takes longer than the delay to shutdown the "restart" can
still act like a "stop".