Fixed dynamic slot ad, it was incomplete
This change uses the ResMgr to more fully compute the attributes on
the dynamically created slot. It is a bit heavy-handed, but appears to
be the best way right now. The Startd's attribute computation code
could use some simplification.

This was found when using a START policy that relied on KeyboardIdle,
like the default UWCS_START does. The policy specifies KeyboardIdle >
15 * 60, but on the dynamic slot the KeyboardIdle was -1 (not yet
initialized). The result was the dynamic slot would always reject
jobs.