Fixed condor_transferer, was hitting get(char *) ASSERT
Back in Oct 2007 (1b017f0eb), the semantics of Stream::get(char *)
were changed. They used to allow for copying the received string into
the buffer argument. That ability was removed and an ASSERT to make
sure the buffer was NULL was put in place. The HAD Utils.cpp
utilSafeGetFile was malloc'ing its own buffer, relying on the old
semantics, and hitting the ASSERT.

This fix changes Utils.cpp to always pass a NULL buffer and free the
result.