Fixed HAD's utilSafeGet/PutFile to not treat an MD5 buffer as if it were a string
The MD5 buffer is not a string and should not be treated like one. It
should not be printed with %s or strcmp'd. It is a fixed size buffer.

This also fixes the fact that when the buffer had a NULL in it only
part of the MD5 would be transferred and compared.