Fixed param(RELEASE_DIR) + /bin -> param(BIN) in HAD
Seriously? param for RELEASE_DIR and then tack /bin onto it?

A further fix for this is to put the condor_transferer in LIBEXEC,
which is where it arguably belongs, param for LIBEXEC and let someone
override the location with a TRANSFERER config macro, like we do for
so many other things.