Restore default port for ec2_sim, with option to specify alternate port
To simplify use of ec2_sim outside of the test suite, it now defaults
to using port 1980 (the port it used to bind to).
But now, it also takes an option -p command-line parameter to specify
an alternate port, which can be 0 (to bind to a dynamic port).
The test suite has ec2_sim bind to a dynamic port to avoid problems
with multiple instances running on the same system.