Rotate HoldReason to LastHoldReason when held jobs are removed
Now, when a held job is removed, the schedd moves HoldReason,
HoldReasonCode, and HoldReasonSubCode to the LastHoldReason and friends.
Otherwise, they can be lost if the job becomes re-held (say if the
gridmanager can't remove the job from the remote host).
Some code in the gridmanager already assumes that HoldReason should only
be defined if the job is held.