Fixed a dirty write bug in the JobRouter's Status Hook
The JobRouter would get a copy of an ad from the ClassAdCollection and
spawn the hook. While the hook was executing the JobRouter would
update the ad the hook was working on. When the hook returned the
stale ad would be used for updates. The result was stale data would
overwrite current data.

The fix is to sync the ad with the ClassAdCollection when the hook
returns.