Fixed gridmanager sending of invalid attribute values to schedd
The amazon_gahp hit an openssl error that had a newline in it. The
error was sent back to the gridmanager who used it as a
HoldReason. The schedd accepted the HoldReason but the transaction log
bailed out because newlines are not allowed in attribute values.

This fix escapes \n's in error messages, which should also make for
nicer logging.