Added invalid attribute value protection (Schedd and client side)
As discovered when the gridmanager wrote a HoldReason with a newline
in it. The Schedd will happily accept attribute values that it cannot
write to the log. A failure that results in an EXCEPT.

This fix introduces AttrList::IsValidAttrValue and uses it to protect
the Schedd when setting attributes, similar to the IsValidAttrName
check. IsValidAttrValue is also used in SetAttribute, along with
IsValidAttrName, to notify clients of the error they are making,
before even contacting the Schedd.