Fixed default RequestMemory = ImageSize -> ceiling(ImageSize/1024.0)
The RequestMemory units are MiB, it is compared directly to a slot's
Memory attribute which is MiB. The ImageSize is in KiB. The old code
was broken such that a 22KiB ImageSize would use up 22MiB on a
slot. Not so bad, but if the ImageSize (calculate initially from
executable size) were say 20MiB it would mean a request for 20GiB of
slot Memory -> failure.