Fix race condition with LocalCredd attribute
This attribute is intended to only be present if the execute
machine can communicate securely with the CredD, and is
required for Windows run-as-owner jobs to match. Previously,
communication would only be tested on startup and reconfig,
so if the CredD were unavailable on startup, the attribute
would be indefinitely absent. This patch causes communication
to be tested periodically.