Ensure we have a sane snprint() on all platforms
On Windows, _snprintf() (which we used to #define to snprintf) won't
null-terminate the buffer if the formatted string is too long.
On HPUX, snprintf() returns -1 if the formatted string is too long.

Now, on all platforms, snprintf() follows the C99 spec. That is, the
buffer is always null-terminated and the return value is the number of
characters that would have been written if the buffere were large
enough.

[v]printf_length() no longer opens /dev/null on each call, except on
HPUX.

Fixed a couple snprintf()/_snprintf() calls on Windows.