Fix setting of ScheddName in gridmanager ads
Contrary to its documentation, the Daemon object won't give you the name
of a daemon if you instantiate the object with a sinful string.
Fortunately, the schedd sets its name in the environment of its
children, which is more reliable anyway. So now we use that when
constructing the gridmanager ads.