Fixed silly sleep 5 in init stop
It was replaced with a call to wait_pid, which waits up to 15 seconds
for the condor_master to exit. An existing issue is if the master
still hasn't exited after those 15 seconds and stop was invoked as
part of restart (or if start is invoked before the condor_master
exists, generally) the subsequent start will think that the master is
already running, which it is, and will not do anything. The result is
the master will eventually exit and not actually be restart.