Prevents startd from crashing if slot types aren't specified.
This is a fix for Red Hat BZ 472599.

Previously, if the user specified a NUM_SLOTS_TYPE_N without
specifying a SLOT_TYPE_N, ResMgr::buildSlot would crash.  This commit
gives AUTO shares of resources to such a slot instead; it does not
warn the user that the slot type is unspecified.

VERSION HISTORY:  Prevents startd from crashing if slot types aren't specified.