Eliminate the CONDOR_EXT, CONDOR_TEST_CNFDTL and CONDOR_TEST_LRG repositories
All of the code in these repos is now either on a public web server (for
tarballs of the external code) or in CONDOR_SRC (everything else). Each
external directory now contains a URLS file that contains a list of URLs
to be fetched before building. This is done by build_external.

The vm universe and ligo tests pull their tarballs from the web server
as part of their setup.

Options to the build system that used to specify a separate location for
the 'externals' directory now only apply to the subdirectories in
'externals' that are written to during the build ('build', 'install',
'triggers').

The vm universe and ligo tests pull their tarballs from the web server
as part of their setup.