This fixes a bug found in ticket #183. Under certain conditions, linux_sig_coredump() became reentrant, and recursed itself to death.
The conditions were if an abort() happened which led into the above signal
handler, and then inside of that handler abort() was called again, it would
reenter and die due to stack overflow. In our case, the abort() was called
inside glibc when it found something it didn't like. Apparently, the
signal handler tickled the exact same bug, which led to the recursion.

The reason this happens is because abort() manhandles the signal mask and
removes SIGABRT from it. So even though we had set up the signal handler
properly with sigaction and a good sa_mask, it didn't matter in the case of
abort().

The fix was to add a SYSV-like reentrancy guard at the top of the function.