Reimplemented and improved HTTP GET support.
The HTTP GET code now supports more MIME types, falls back to sensible
behavior in the event of certain failures that were previously fatal,
allows files to be accessed from within subdirectories, and checks to
ensure that the requested file is in a subdirectory of WEB_ROOT_DIR.

Notable changes:

0.  The handler code was completely reimplemented; in so doing, it was
also substantially simplified and streamlined.

1.  The GET handler now has explicit support for serving many
more kinds of files with the correct MIME types.

2.  The GET handler will no longer fail with a 403 error if a client
is attempting to access a file with an unsupported MIME type.

3.  The GET handler now allows clients to access files stored in
subdirectories of WEB_ROOT_DIR and has a more robust mechanism (based
on realpath) to ensure that malicious requests cannot access files
outside of WEB_ROOT_DIR.  Client attempts to browse directories are
met with a 403.

VERSION HISTORY:  Reimplemented and improved HTTP GET support for static files in WEB_ROOT_DIR.