Made dollarDollarExpand() more efficient.
Whether to store expansion info persistently or not may now be optionally
turned off when calling dollarDollarExpand().  It is turned off in all cases
except when getting the ad to send to the shadow, because that is the only
case where we need to be able to restore the expansions after a restart.

Also, a single transaction is now used for the whole ad, rather than two
transactions per attribute being expanded.