Workaround for uninitialized memory access bug in SOAP, around transaction termination (commit and abort).
The Schedd has a global Q_SOCK pointer to a QmgmtPeer. The SOAP level
owns the memory the Q_SOCK points to, and the Schedd tries to use the
Q_SOCK after the memory has been delete'd. This fix is to clear out
the Q_SOCK before free'ing the memory it points to.

VERSION HISTORY: see #310