The dprintf error log was getting clobbered in some cases.
When a dprintf error log is created, a subsequent failure in
dprintf_unlock() was overwriting the original dprintf error log
with a new error message about the unlock failure. Since
dprintf_unlock() is being called without even checking to see
if we are in the locked state, the unlock "error" might simply
be that we never successfully locked in the first place. We now
only update the dprintf error log once.