Gridmanager wouldn't send a gt2 cleanup signal in some failure cases.
When restarting a jobmanager, the GlobusJob object wouldn't note that
it had job status callbacks registered with the new jobmanager process.
As a result, after any subsequent cancel command, GlobusJob wouldn't
wait for a callback response or send a commit end signal (needed for
the jobmanager to do job cleanup).