Made RequestMemory's default consider JobVM_Memory before ImageSize
The RequestMemory attribute is supposed to be an estimate of a job's
size to help in construction of a dynamic slot. If it isn't explicitly
given it defaulted to the ImageSize. However, in the case of a VM job
there is also a good estimate of memory from the JobVM_Memory
attribute. This commit makes RequestMemory use vm_memory and fall back
to image_size