Moved SCHEDD_ADDRESS_FILE and SCHEDD_DAEMON_AD_FILE from LOG to SPOOL
These two files are dropped by the schedd and are used by local tools,
and Quill, to locate the Schedd without contacting the Collector. Right
now they default to -

SCHEDD_ADDRESS_FILE  = $(LOG)/.schedd_address
SCHEDD_DAEMON_AD_FILE = $(LOG)/.schedd_classad

This is all well and good, except if you are in an HA setup. When you
have fail-over of the schedd you'll get stale files on the failed schedd
machine. From that point forward tools, e.g. condor_q/submit, on the
failed schedd machine will not be able to contact the schedd. The tools
consult the files and do not fall back to a collector lookup.

Solutions? 1) make the tools fall back to a collector lookup, 2) don't
use the files at all if you are in an HA schedd setup, 3) put the files
in SPOOL instead of LOG

(1) is work with little payoff at the moment
(2) works, but requires separate configuration when in HA mode
(3) avoids the work of (1), allows for a consistent config over (2), and
may even benefit from letting multiple machines in a pool avoid the
collector lookup

Downsides of (3)? Well, the file has been in $(LOG) for a long time,
along with all other ADDRESS_FILEs, but no one should be relying on that!