Fixed units on vm_memory in the libvirt vm gahp
condor_submit's manual page says vm_memory is in MiB, the VM_MEMORY
configuration for the vm gahp says it is in MiB, and libvirt says that
the <memory> element of the VM's <domain> is in KiB.

The vm-gahp was passing vm_memory through to libvirt without doing the
MiB->KiB conversion.

VERSION HISTORY: Fixed vm-gahp's handling of a job's vm_memory, there was a unit mismatch.