Improve how the grid monitor invokes globus-url-copy and inner script
* Continue to send OK messages while globus-url-copy is running.
* Kill globus-url-copy when timeout occurs.
* Guarantee we have the inner script's pid stored before it can be
  reaped.

During testing, we found that when fork()/exec()ing a child process, if
the child exits quickly enough, the parent's SIGCHLD handler is called
before fork() returns in the parent. Thus, we end up reaping the child
process before we know its pid. Our reaper needs to know the pids of the
inner script and globus-url-copy. Thus, we now have safe_fork(), which
blocks SIGCHLD until the pid of the child is saved in a variable.

gittrac #429