Fixed SEGV during startd shutdown when a dynamic slot was waiting on work fetch
During shutdown, slots are put into the unclaimed state and then
deleted. A dynamic slot will normally be deleted as soon as it hits
the unclaimed state, and will return its resources to its parent. This
is a harmless activity during shutdown.

A dynamic slot that has an outstanding fetch work hook, will not
delete itself as soon as it hits the unclaimed state. It will defer
its destruction. However, during shutdown all slots are forcibly
deleted. During that deletion they try to return their resources to
their parent, except it is possible their parent was already deleted
resulting in a SEGV.

The solution here was to recognize that if a dynamic slot is being
deleted and it has outstanding hooks then the startd is being shutdown
and there is no point in returning resources to a parent.