Revert the revert of buffered i/o in quill's job queue reader.
The underlying problem hit by the job_quill_basic test has been understood
and fixed.  It was not caused by a bug in this patch.  It was caused by
changes in the timing of events, which changed the result of a race
condition. #450

Revert "Revert "Got rid of unbuffered i/o implementation in quill's ClassAdLogParser.""

This reverts commit 5d926cc883008e56dc29fd6f0a9518c91848f4c3.