Fixed shutdown race condition in master with HA daemons
HA daemons in the master would never get notified that they were
shutting down, because they did not have a pid yet. The result was
during a shutdown, an HA daemon could acquire its lock and start
running. Once running it would prevent the master from completely
shutting down (a good thing). This was further problematic because the
master would ignore the signal it had received for shutdown. A
workaround was to send a different signal, e.g. if first sent QUIT
second time send TERM.